trash/foresight/requirements.txt
numpy
celery
channels
channels-redis
Django
django-crispy-forms
django-filter
pandas
psycopg2-binary
pymongo
redis
requests
xlsxwriter
scikit-learn
django-import-export
sqlalchemy
paho-mqtt
scipy